Adds unitest for unsupported YAML constructor
Moves generic pytest methos into test_cwd/utils.py
