+++ tools/kcli/tests/test_cwd/IRYAMLConstructorError.yml
@@ -1,0 +1,3 @@
+---
+key:
+  sub_key: !notag
