--- tools/kcli/tests/test_cwd/consts.py
@@ -1,5 +1,0 @@
-import os
-
-TESTS_CWD = os.path.dirname(__file__)
-SETTINGS_PATH = os.path.join(TESTS_CWD, "settings")
-# CONFIG_FILE = os.path.join(TESTS_CWD, conf.KCLI_CONF_FILE)
