+++ tools/kcli/tests/test_yaml.py
@@ -1,0 +1,17 @@
+import os.path
+
+import pytest
+import configure
+
+from tests.test_cwd import utils
+
+our_cwd_setup = utils.our_cwd_setup
+
+
+def test_unsupported_yaml_constructor(our_cwd_setup):
+    from kcli.utils import update_settings
+    from kcli.exceptions import IRYAMLConstructorError
+    tester_file = 'IRYAMLConstructorError.yml'
+    settings = configure.Configuration.from_dict({})
+    with pytest.raises(IRYAMLConstructorError):
+        update_settings(settings, os.path.join(utils.TESTS_CWD, tester_file))
